Change smoothing logic to run at/below thresholds
For finishing it's more natural to specify levels where you
want to be running a finishing op and hence above that level
is the semi finishing, etc. So change the tests for levels
and adjust comments for clarity.

Similarly for roughing it is more intuitive to have a level
that triggers full roughing mode