Per op accuracy mode override, improve accuracy mapping
We add a new dropdown in the CAM workspace for all ops
allowing override of the accuracy mode for that op

NOTE: This override takes preference OVER what is set
in the post screen params

NOTE: We shuffle all param values up by 1, leading to a
problem when reposting old programs. Dialog will be showing
different accuracy by 1. However, the old post had a bug
leading to an "off by 1" (on M298), meaning it was writing
the value 1 different to what you selected in the list...
So post page will change from what is expected, but actually
will post the same output as previously (ie it was wrong before!)

Update old post settings if not using "Automatic"

M298 modes are changed based on some testing.
L1/L4 seem to be the stars from a speed point of view.

Avoid using "smoothing" in Fusion for any finish ops
(ie M298 L5/6). These are handled very badly by the C00
controller leading to jerky and slow output. Send unsmoothed.