Change smoothing logic to run at/below thresholds
For finishing it's more natural to specify levels where you
want to be running a finishing op and hence above that level
is the semi finishing, etc. So change the tests for levels
and adjust comments for clarity.

Change smoothing threshold calculation

Change calculation of smoothing threshold to not require BOTH
vertical and horizontal stock to leave before a given roughing
threshold is triggered. For rough and semi-rough we require SOME
horizontal stock to leave, but otherwise either our rule is either
horizontal OR vertical stock to leave will trigger that roughing level.

This allows full roughing where the vertical stock to leave might be
negative, eg around stock perimeter

Similarly for roughing it is more intuitive to have a level
that triggers full roughing mode