G100 handles stopping spindle and setting coolant
G100 can handle repositioning, coolant changes and
stopping/restarting the spindle.

Saves about 0.2s to let G100 stop the spindle

Letting it restart the coolant gets it flowing *just* in time before
tool is back in the cut. Really needs a small dwell to be sure.

There is no reason to do a manual retract, G100 will handle this