G68 rotation: Linearise moves and suppress change of G17/18/19
Brother doesn't allow calling G17/18/19 while G68 rotation is active
This means we need to linearise any G18/19 lead-in moves

Also the initial G17 call in OnSection() happens before we release
the G69 rotation. So add a G69 at start of OnSection(), we will
reset G68 later in the section.