Optionally dog-leg G100 toolchange
Some machines (R, F) have a fixed tool change location,
so the G100 toolchange isn't so useful for these machines.
In particular it causes a direct movement from toolchange
location to part.

We add an option to allow the user to choose a
"dog leg" move where we move first in XY, then in Z